fix(lightningcss): add scss/sass support
