+++ packages/lune/src/builder/esbuild/index.ts
@@ -1,37 +1,8 @@
-import type { Plugin } from "esbuild";
 import { build } from "esbuild";
 import { solidPlugin } from "esbuild-plugin-solid";
 import { sassPlugin, postcssModules } from "esbuild-sass-plugin-ysink";
-import { LuneCfg } from "../config";
-
-const ShelterSolidResolver = (): Plugin => {
-  const resolverRoots = {
-    "solid-js": "solid",
-  };
-  return {
-    name: "shelter-solid-resolver",
-    setup(build) {
-      build.onResolve({ filter: /solid-js(?:\/web|\/store)?/ }, ({ path }) => ({
-        path,
-        namespace: "shltr-res-ns",
-      }));
-      build.onLoad({ filter: /.*/, namespace: "shltr-res-ns" }, ({ path }) => {
-        const pathSplit = path.split("/");
-        const resolvedPath =
-          resolverRoots[pathSplit[0]] +
-          pathSplit
-            .slice(1)
-            .map((s) => s[0].toUpperCase() + s.slice(1))
-            .join("");
-
-        return {
-          contents: `module.exports = shelter.${resolvedPath}`,
-          loader: "js",
-        };
-      });
-    },
-  };
-};
+import { LuneCfg } from "../../config";
+import { ShelterSolidResolver } from "./resolver";
 
 export async function createEsbuildBuilder(entryPoint: string, outfile: string, minify: boolean, cfg: LuneCfg) {
   if (cfg.builder && cfg.builder !== "esbuild")
